FastApi_3/Task 6.5/requirements.txt
fastapi
uvicorn
python-dotenv
pyjwt
passlib[bcrypt]
slowapi